video-ai-system/scripts/requirements.txt
# It's recommended to use a specific version of mmcv-full that matches your PyTorch and CUDA versions.
# This example is for torch 1.12.0 and cuda 11.3, as seen in the notebook.
# Users should check https://download.openmmlab.com/mmcv/dist/ for the correct version.
mmcv==2.1.0
mmaction2==0.25.0
timm
einops
decord